New feature: settings
New dependency: electron-settings
New class: VineEditorSettings, it handles electron-settings for us. Only one instance per program.

The app now creates a settings file on launch if one doesn't already exists.
Configurations, user preferences and data are read from it and used to setup the window, define the type of node connection, etc.
'vine-editor-settings.js' contains a template from which default values are taken.
Writing to the settings file is done by default when the app is closed.
I'm planning to add the preferences window later on to make it easier for users.